Minor fixes to allow usage.
Added 'gzip' option to score annotators to indicate whether the score
files are in compressed format or not.

If no score config is passed to the missense_scores annotator, it will
search for a score config file with the same name as the score file.

If the default search columns 'ref' and 'alt' in the missense_scores
annotator are not found in the input file's header, they will be
omitted.

Changed the FreqTransm annotator's frequency columns' delimiter from '.'
to '_'.

annotate_with_multiple_scores now uses its own new columns to determine
the order of the new columns.

Fixed header string reading in utilities.py

Adjusted tests accordingly.