Added file collection via wildcard. Updated tests.
Removed default current chrom value of '1' and added a method to list
all files in the dbNSFP directory that match the given dbNSFP wildcard.
The dbNSFP annotator is then initialized with any one of these files.

Updated unit tests to be more concise, annotate with two columns and use
indel variants.