Add fetch_region to PositionScore. Use it to build histograms.
fetch_region returns a dictionary of generators. Each generator will
yield the scores for the corresponding score_id. The generators may
yield not just the values in a range but all scores in the entire
resource.

This functionality is used to build a HistogramBuilder class which takes
a resource and build histograms according to it's configuration.