Fix fetch_region API to prevent multiple iterations over input file
fetch_region used to return a dict of generators. Each generator would
generate score for a specific score_id. This ment we would iterate over
the input file multiple times, once for each score.

The new API returns a generator of dicts. Each dict has all the values
for all the scores for a line in the input file.