Investigate missing ref allele frequency after deserialization
The reference allele is usually not serialized and stored in the DB.
This allele is implicit and its values are calculated from the other
alleles when a variant is deserialized. This recalculation was
incomplete, it was leaving count and frequency for the ref allele
uncalculated (Nones).

Fixing this for schema2 is trivial, fixing it for schema1 is not. It
would require a change in the serialization format used by schema1 and
this in turn would invalidate currently imported data. So we decided not
to do it. Instead schema2 follows schema1 erronious behaviour. This
should be fixed onces schema1 is removed.

Unit and integration tests (for both schema1 and schema2) that cover
this case have been disabled.