Set input files of a graph to config files of a project
No need to manually do that. It is automatically done by import-tools.
As a result of this change tasks may get unnecesary recalculated but
this may also save us a lot of problems when a config file's input
params get changed but not the spliting into buckets or task ids.