Add count_measures to remote phenotype studies
count_measures is an abstract method in the PhenotypeData
base class and is used in the pheno browser download.
This prevented remote phenotype studies from being created, because
they lacked the count_measures implementation.

A new API view for providing the return value of count_measures
is added and a getter is added to the REST client. This is used
in the remote phenotype study to implement count_measures.