Refactor annotate_schema2_parquet test parametrization
The unit tests for annotate_schema2_parquet must be run for different
types of input data, which is done using parametrization. Input data is
provided via pytest fixtures and parametrization with different fixtures
is not immediately obvious. This led to a sub-par initial
implementation, which this commit rectifies. It also adds one more study
to the parametrized data, which obsoleted two of the unit tests written
separately.