Implement instance reannotation tool
This commit adds the initial implementation of an instance reannotation
tool, which will walk through all configured studies and perform
reannotation on them. Currently, it works only with duckdb_parquet and
parquet storage types. Further work is needed to detect if a study's
annotation is up-to-date and outright skip it if necessary, as well as
to implement reannotation for other storage types.