Change annotation tools work dir setup behaviour
Instead of setting up/creating the work dir for TaskGraph logs in the
init method of the tool, set it up when actually running the tool. This
change was done to allow implementing separate behaviour for the
AnnotateSchema2Parquet tool which does not require a TaskGraph,
rendering the creation of a work dir unnecessary.