Update reference import to import multiple instruments
The reference import was implemented for benchmarking against duckdb
with support for only a single instrument file.
After deciding to use the reference implementation in the full import,
it is now updated to support importing an entire instrument directory.
The implementation creates a task graph with 1 task for every measure in
every instrument. The task opens and reads the instrument file,
recording and classifying a specific measure.