Fix TabixGenomicPositionTable behaviour for relabeled contigs
The issue was caused due to erroneous use of the "raw" (i.e.
non-relabeled) contig name instead of the adjusted (i.e. relabeled)
contig name. Equality checks for the contig name would fail even though
the contig was in fact the same, leading to the buffering mechanism
being bypassed. This was the cause for massive slowdowns when annotating
with Tabix-based scores that utilised contig relabeling.
Variable names have been standardized to use "chrom" when
referencing relabeled/canonical variable names and "fchrom" for raw,
non-relabeled contig names.
Usage of "fchrom" when querying was confined to the method interfacing
with Tabix.