fix gene sets denovo datasets hierarchy
* The hierarchy in modal contains all datasets
but it should show only the visible ones. A check if
the dataset or study is visible is added when creating the hierarchy
for the modal. Visible datasets are provided by DatasetsService.

* It is possible while constructing the hierarchy
a node to be null, so a check is added to prevent
saving null value to the array which stores
the hierarchy.

* Add unit test for the case when a node is null.
* Mock datasets service method which provides the visible datasets
 and update hierarchy mock to cover the case of a dataset which has
 children but they aer not visible.