Disable default writing of reference alleles in VariantsParquetWriter
This was done to match the default behaviour of import tools, as before
this change, reannotating a study without reference alleles would
include them, increasing the amount of data written. Future improvement
will be done to store this flag in the metadata of Parquet studies and
read it when reannotating, to determine whether to include reference
alleles.