Add pheno_common instrument to pheno import
When changing to the new implementation, this was leftover and is now
added back. The new handling reads the pedigree as an instrument file
and adds every default pedigree column as a skip in the inference
configuration.
Due to strange pedigree column conventions, the columns are now hard
coded, because the ones in the PEDIGREE_COLUMN_NAMES variable do not
match the default ones in the pedigree import, and should be fixed
later.