Fix handling of frequency filter when adding frequency bins
We had a bug in handling frequency filters in
`calc_frequency_bins` method.

The problem was that if the frequency filter is not related to
`af_allele_freq`, we added the denovo frequency bin even when
it is not needed.

In the gene browser, we always pass a frequency filter, based
on gnomAD frequencies, and it resulted in adding frequency_bin={0}
to the query.

Now, we first check if the frequency filter is related to
`af_allele_freq`, and only then can we manipulate the frequency
bins.