Refactor LineBase to act as a simple protocol instead of a parent class
LineBase was set up as an ABC and classes inheriting from it were
calling LineBase's constructor. This was unnecessary, as LineBase's
constructor did nothing, and it only slowed down instantiation of these
classes. To remedy this, super calls to __init__ were removed and
LineBase was transformed into a simple Protocol type.