Fix maximally_extend_variant infinite loop
While importing the SSC WGS 2 study, we encountered and
infinite loop in the `maximally_extend_variant` utility function
used in the liftover annotator.

The problem was that after the end of a chromosome, the reference
genome get_sequence() method would return an empty string.
When extending the ref and alt beyond the chromosome end,
we get an empty string, append it to the end of the ref and alt
and believe that this changes them.

Now we check that when extending ref and alt, we actually
change them.