Removing genomic resource NPScore
While evolving the interface of fetch_region and fetch_scores
we have realized that the AlleleScore and NPScore are representing
the same type of resource.

We have started working on the removal of the NPScore genomic score.

We still allow usage of `np_score` resource type in the resource
configuration but the class created is AlleleScore.