NPScoreAnnotator removed
We are merging np_score with allele_score. The only difference
between these two socres was handling of short indels. We
introduced a parameter in the allele_score resource named
`substitutions_only`. For missense scores or scores similar to
CADD this parameters should be set to True. For frequencies
like gnomAD, this parameter should be set to False.