Correctly produce Tabix index files in annotate_columns
Previously, when producing an output tabix file, we'd assume the indices
of the chrom, position and end position columns; however this does not
always reflect the input data and would fail for input with non-standard
columns. This commit fixes the produce_tabix_index function to always
collect the actual indices for the relevant columns, instead of assuming
them.