Optimization for measure classification in pheno import
Previously, a Dask task would be created for each measure to be
classified. This was costing time as the instrument files were being
read many times over and over. This commit changes the creation of tasks
for classification so that each task will handle a whole instrument
instead of a single measure.