Genotypes in tandem repeats could be large
We used np.int8 for storing family genotypes. Since we moved
to newer version of NumPy this causes an exception for some
tandem repeat variants.

We switched family genotypes to np.int64.

For Schema1 we still insist that genotypes can be stored in
signed np.int8 because of the serialization used.