Always merge person set collections for data groups
Fixes bug where a genotype data group with a single study would skip
merging its person set collection configuration. This would skip
updating the values for the constituent person sets; if this genotype
data group was used as a child for another genotype data group, it would
not be able to properly load its person sets since the values will not
reflect those in the genotype data group's pedigree.