Fix handling of user's default group name
The user's default group name conicides with the user email
addres. We changed the behaviour of the GroupsViewSet to accept
'.' as valid character in the name of a group.

Default user fixture is changed to create the default user's
group for user 'user@example.com'.

Test for user's group permission is added. Some tests are updated.